moving modules into where they need to go.
